Update comments for PR
